Tighten hook type checking
If a hook is successful, it must return a task.  If a hook isn't
successful, it cannot return a task or a post-hook action, and must
return a message.

This still isn't as tight as it could be, as it'll allow a successful
return code with no task, provided there's a message and no post-hook
action.  Fixing that would require using one of the typing extensions
that can exclude literal values, for example using the annotated-types
library.