hooks: add fix_weekday_due
It turns out that tasks that are set to be due on weekdays at 00:00 end
up sometimes being due on Sunday at 23:59:59.  That's not what I want!
Add a hook to fix this up by detecting tasks that have ended up as due
on a Sunday and recur on weekdays, and correcting them.